requirements.txt
# Core dependencies
numpy>=1.20.0
matplotlib>=3.4.0
Pillow>=8.2.0
reportlab>=3.6.0

# PDF generation dependencies
PyPDF2>=2.0.0

# Path handling
pathlib>=1.0.1
